2YMC_1차 프로젝트/1.OCR_AI_프로젝트/D.구현및보고/2_MVP Prototype/requirements.txt
streamlit
python-dotenv
requests
pillow
opencv-python
openai>=1.42.0
pandas